Simplified active agenda install procedure.
The installation happens late in the packaging process, so that the chrome.manifest doesn't interfere with the omni.ja creation. However, Adrien was already sending the files in a way that they could simply be copied into a regular firefox build, so it makes sense to do exactly that during the build. The new structure of /active-agenda reflects this.